Blob data is now stored in nested directories.
Files are stored by their hashes (given in hex)
In order to keep the number of files per directory managable, the first
two bytes of the hash are stripped and used as the names of two nested
directories into which the file is placed.
Ex. hash = 3bc6e94a89
	 => relpath = 3b/c6/e94a89

Reworked storing data so that relative file paths and byte are offsets
are returned from IBlobStoreDependencies.StoreBlob() not inputs.