BlobLocations no longer includes BlobType.
Also removes IsMultiblock reference.

Together removes dependence of BlobLocation on what the data is that it
stores. This information must now be provided when interpreting data
read from the blob store. Fixes possible problems occurring with a
single blob of data representing two differnt things. For instance a
part of a split file could conceivable be backed up as a single file. Or
a MetadataNode could be backed up then transferred to the source folder
to be backed up as its own file.