Fixed not adding file references during backup.
For files being backed up a second time that were identified as containing
metadata only changes or were unchanged, the blobs representing were not
referenced again. Thus if the first backup was deleted, the file data no
longer existed in the backup store despite the second backup still
containing references to that data.