Multiblock references are now only stored in the blob index.
They are no longer stored as their own blobs. This is more efficient,
and since the average blob size should be ~4MB, the size of the hash
list needed to store a large file is 1/4e6 * 20(hash len) = 1/200,000th
the size of the file in question.

Modified dict_decode so the objects with no length are returned as bytes
of length 0, not null. This produces the desired behavior for strings:
saving (mykey: "") -> string encode -> (mykey: byte[0]) -> dict_encode
-> dict_decode -> (mykey: byte[0]) -> string decode -> (mykey: "")