Fixed an issue with TransferBackup and an issue with RunBackup().
TransferBackup would previously read blobs with a BlobReferenceIterator immediately after writing them.
When run on a disk, the just written blob might not be available, even though the write had completed.
Now blobs are kept in memory immediately after write and passed to the iterator. Additionally the iterator now works from the transfer source not the destination.

Fixed a crash that would occur when adding a new destination after a backup, then running another backup, due to a bug in RunBackup itself.